Guard against crashes when canonical-based URLs are invalid
Processing code assumed that canonical-based URLs would have multiple parts after the canonical base.  When resources mistakenly didn't follow this pattern, the publisher would crash.  Now it checks the length of the split array before trying to index into it.  The invalid URL is then captured and logged further down the processing chain and correctly flagged as an invalid URL.